requirements.txt
fastapi==0.104.1
pydantic==2.4.2
starlette==0.27.0
passlib==1.7.4
sqlalchemy==2.0.22
alembic==1.12.1
boto3==1.28.74
uvicorn[standard]
python-jose[cryptography]

pymysql==1.1.0
# .env
python-dotenv==1.0.0
python-decouple==3.8
pydantic-settings==2.0.3

